requirements.txt
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.7.1
seaborn>=0.12.2
jinja2>=3.0.0
pyyaml>=6.0.0
google-cloud-bigquery>=3.0.0
google-cloud-bigquery-storage>=2.19.1
pandas-gbq>=0.19.0
pyarrow>=12.0.0